Refactor pot2 function for cleaner code

+++ Tarefa de Laboratório 05/Q6.c
@@ -3,7 +3,7 @@
 double pot2(int n) {
     if (n<0) return 1./pot2(-n);
     if (n == 0) return 1.;
-    return (double) 2.*pot2(n-1);
+    return 2.*pot2(n-1);
 }
 
 int main() {
